+++ src/MyBirdTap/mocks/bugArray.ts
@@ -1,0 +1,5 @@
+export const bugArray: ('morningBug' | 'nightBug' | 'goldenBug')[] = [
+  'morningBug',
+  'nightBug',
+  'goldenBug'
+];
