+++ src/MyBirdTap/mocks/bugColor.ts
@@ -1,0 +1,7 @@
+import { BugColor } from './types/bugColor';
+
+export const bugColor: BugColor = {
+  morningBug: 'text-light-point',
+  nightBug: 'text-dark-point',
+  goldenBug: 'text-warning'
+};
