+++ src/MyBirdTap/mocks/types/birdItem.d.ts
@@ -1,0 +1,10 @@
+export interface BirdItemType {
+  id: string;
+  type: 'MORNING' | 'NIGHT';
+  category: string;
+  name: string;
+  image: string;
+  level: number;
+  bugPrice: number;
+  goldenBugPrice: number;
+}
