+++ src/MyBirdTap/mocks/types/birdItems.d.ts
@@ -1,0 +1,12 @@
+import { BirdItemType } from './birdItem';
+
+export interface BirdItemsType {
+  MORNING: {
+    purchasedItems: BirdItemType[];
+    notPurchasedItems: BirdItemType[];
+  };
+  NIGHT: {
+    purchasedItems: BirdItemType[];
+    notPurchasedItems: BirdItemType[];
+  };
+}
