+++ src/MyBirdTap/mocks/types/bugColor.d.ts
@@ -1,0 +1,5 @@
+export interface BugColor {
+  morningBug: string;
+  nightBug: string;
+  goldenBug: string;
+}
