+++ src/MyBirdTap/mocks/types/wallet.d.ts
@@ -1,0 +1,5 @@
+export interface Wallet {
+  morningBug: number;
+  nightBug: number;
+  goldenBug: number;
+}
