+++ src/MyBirdTap/mocks/wallet.ts
@@ -1,0 +1,7 @@
+import { Wallet } from './types/wallet';
+
+export const wallet: Wallet = {
+  morningBug: 4,
+  nightBug: 300,
+  goldenBug: 100
+};
